Use the newly introduced IndexFilter in milli
