+++ crates/milli/src/search/facet/mod.rs
@@ -5,7 +5,7 @@ use heed::{BytesDecode, RoTxn};
 use roaring::RoaringBitmap;
 
 pub use self::facet_distribution::{FacetDistribution, OrderBy, DEFAULT_VALUES_PER_FACET};
-pub use self::filter::{BadGeoError, Filter, SHARD_FIELD};
+pub use self::filter::{BadGeoError, Filter, IndexFilter, SHARD_FIELD};
 pub use self::search::{FacetValueHit, SearchForFacetValues};
 use crate::heed_codec::facet::{FacetGroupKeyCodec, OrderedF64Codec};
 use crate::heed_codec::BytesRefCodec;
@@ -17,7 +17,6 @@ mod facet_range_search;
 mod facet_sort_ascending;
 mod facet_sort_descending;
 mod filter;
-mod filter_vector;
 mod search;
 
 fn facet_extreme_value<'t>(
