+++ crates/milli/src/test_index.rs
@@ -13,6 +13,7 @@ use crate::constants::RESERVED_GEO_FIELD_NAME;
 use crate::error::{Error, InternalError};
 use crate::index::{DEFAULT_MIN_WORD_LEN_ONE_TYPO, DEFAULT_MIN_WORD_LEN_TWO_TYPOS};
 use crate::progress::Progress;
+use crate::search::facet::IndexFilter;
 use crate::update::new::indexer;
 use crate::update::settings::InnerIndexSettings;
 use crate::update::{
@@ -497,56 +498,72 @@ fn test_basic_geo_bounding_box() {
 
     // exact match a document
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([0, 0], [0, 0])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([0, 0], [0, 0])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[0]>");
 
     // match a document in the middle of the rectangle
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([10, 10], [-10, -10])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([10, 10], [-10, -10])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[0]>");
 
     // select everything
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([90, 180], [-90, -180])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([90, 180], [-90, -180])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[0, 1, 2, 3, 4]>");
 
     // go on the edge of the longitude
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([0, -170], [0, 180])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([0, -170], [0, 180])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[1]>");
 
     // go on the other edge of the longitude
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([0, -180], [0, 170])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([0, -180], [0, 170])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[2]>");
 
     // wrap around the longitude
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([0, -170], [0, 170])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([0, -170], [0, 170])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[1, 2]>");
 
     // go on the edge of the latitude
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([90, 0], [80, 0])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([90, 0], [80, 0])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[3]>");
 
     // go on the edge of the latitude
     let search_result = search
-        .filter(Filter::from_str("_geoBoundingBox([-80, 0], [-90, 0])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([-80, 0], [-90, 0])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[4]>");
@@ -555,7 +572,9 @@ fn test_basic_geo_bounding_box() {
 
     // try to wrap around the latitude
     let error = search
-        .filter(Filter::from_str("_geoBoundingBox([-80, 0], [80, 0])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([-80, 0], [80, 0])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap_err();
     insta::assert_snapshot!(
@@ -568,7 +587,9 @@ fn test_basic_geo_bounding_box() {
 
     // send a top latitude lower than the bottow latitude
     let error = search
-        .filter(Filter::from_str("_geoBoundingBox([-10, 0], [10, 0])").unwrap().unwrap())
+        .filter(IndexFilter::from(
+            Filter::from_str("_geoBoundingBox([-10, 0], [10, 0])").unwrap().unwrap(),
+        ))
         .execute()
         .unwrap_err();
     insta::assert_snapshot!(
@@ -604,17 +625,19 @@ fn test_contains() {
     let rtxn = index.read_txn().unwrap();
     let mut search = index.search(&rtxn);
     let search_result = search
-        .filter(Filter::from_str("doggo CONTAINS kefir").unwrap().unwrap())
+        .filter(IndexFilter::from(Filter::from_str("doggo CONTAINS kefir").unwrap().unwrap()))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[0, 1]>");
     let mut search = index.search(&rtxn);
-    let search_result =
-        search.filter(Filter::from_str("doggo CONTAINS KEF").unwrap().unwrap()).execute().unwrap();
+    let search_result = search
+        .filter(IndexFilter::from(Filter::from_str("doggo CONTAINS KEF").unwrap().unwrap()))
+        .execute()
+        .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[0, 1, 2]>");
     let mut search = index.search(&rtxn);
     let search_result = search
-        .filter(Filter::from_str("doggo NOT CONTAINS fir").unwrap().unwrap())
+        .filter(IndexFilter::from(Filter::from_str("doggo NOT CONTAINS fir").unwrap().unwrap()))
         .execute()
         .unwrap();
     insta::assert_debug_snapshot!(search_result.candidates, @"RoaringBitmap<[2, 4, 5]>");
@@ -1383,9 +1406,10 @@ fn vectors_are_never_indexed_as_searchable_or_filterable() {
     assert!(results.candidates.is_empty());
 
     let mut search = index.search(&rtxn);
-    let results =
-        dbg!(search.filter(Filter::from_str("_vectors.doggo = 6789").unwrap().unwrap()).execute())
-            .unwrap();
+    let results = dbg!(search
+        .filter(IndexFilter::from(Filter::from_str("_vectors.doggo = 6789").unwrap().unwrap()))
+        .execute())
+    .unwrap();
     assert!(results.candidates.is_empty());
 
     index
@@ -1415,7 +1439,7 @@ fn vectors_are_never_indexed_as_searchable_or_filterable() {
 
     let mut search = index.search(&rtxn);
     let results = search
-        .filter(Filter::from_str("_vectors.doggo = 6789").unwrap().unwrap())
+        .filter(IndexFilter::from(Filter::from_str("_vectors.doggo = 6789").unwrap().unwrap()))
         .execute()
         .unwrap();
     assert!(results.candidates.is_empty());
